backend/requirements.txt
fastapi
fastapi_cache2
python-jose
loguru
passlib
psycopg2_binary
sse-starlette
pydantic
pydantic[email]
pydantic_settings
python-dotenv
redis
Requests
SQLAlchemy
uvicorn
python-multipart
alembic
bcrypt